pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "misinfo-moderation-env"
version = "1.0.0"
description = "OpenEnv-compliant RL environment for AI-driven misinformation moderation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black", "isort"]

[tool.setuptools.packages.find]
where = ["."]
